Controllers are extracted to the controller package
