+++ Communitter/src/main/java/com/communitter/api/controller/CommunityController.java
@@ -1,5 +1,8 @@
-package com.communitter.api.community;
+package com.communitter.api.controller;
 
+import com.communitter.api.community.Community;
+import com.communitter.api.community.CommunityService;
+import com.communitter.api.community.Subscription;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
