+++ Communitter/src/main/java/com/communitter/api/controller/PostController.java
@@ -1,6 +1,8 @@
-package com.communitter.api.post;
+package com.communitter.api.controller;
 
 
+import com.communitter.api.post.Post;
+import com.communitter.api.post.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
